Make random number generator produce 64-bit values.
Return a 64-bit value rather than 32-bit value. No other functional
difference.

Note that the random number generator here is very simple, and can not
be relied on for making good random numbers.

+++ runtime/math.c
@@ -13,7 +13,7 @@ static uint64_t rand_seed = 8919118912341193UL;
 static const unsigned rand_mul = 1103515245U;
 static const unsigned rand_add = 12345;
 
-static unsigned urand()
+static uint64_t urand()
 {
     rand_seed = rand_mul * rand_seed + rand_add;
     return rand_seed;
